+++ lib/presentation/screens/workout/workout_01_process_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:npng/logic/cubit/workout_cubit.dart';
+import 'package:npng/widgets/help_icon_button.dart';
 import 'package:wakelock/wakelock.dart';
 
 import 'package:npng/data/models/models.dart';
@@ -22,6 +23,7 @@ class WorkoutProcessScreen extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(
         title: Text(S.of(context).currentWorkout),
+        actions: [HelpIconButton(help: S.of(context).hintWorkoutProcess)],
       ),
       persistentFooterButtons: [
         Builder(